Split health checks into separate file
